[Modify] ArticleCreateScreen.tsx Update
